+++ Sources/Sentry/SentryNSProcessInfoWrapper.mm
@@ -2,6 +2,14 @@
 
 @implementation SentryNSProcessInfoWrapper
 
++ (SentryNSProcessInfoWrapper *)shared
+{
+    static SentryNSProcessInfoWrapper *instance = nil;
+    static dispatch_once_t onceToken;
+    dispatch_once(&onceToken, ^{ instance = [[self alloc] init]; });
+    return instance;
+}
+
 - (NSString *)processDirectoryPath
 {
     return NSBundle.mainBundle.bundlePath;