+++ Sources/Sentry/include/HybridPublic/PrivateSentrySDKOnly.h
@@ -64,6 +64,11 @@ typedef void (^SentryOnAppStartMeasurementAvailable)(
  */
 + (NSString *)getSdkVersionString;
 
+/**
+ * Retrieves extra context
+ */
++ (NSDictionary *)getExtraContext;
+
 @property (class, nullable, nonatomic, copy)
     SentryOnAppStartMeasurementAvailable onAppStartMeasurementAvailable;