+++ Tests/SentryTests/Helper/SentrySerializationTests.swift
@@ -12,18 +12,7 @@ class SentrySerializationTests: XCTestCase {
             "valid object": "hi, i'm a valid object",
             "invalid object": NSDate()
         ]
-
-        var data: Data?
-        let exp = expectation(description: "Should throw error")
-        do {
-            data = try SentrySerialization.data(withJSONObject: json)
-        } catch {
-            exp.fulfill()
-            //Depending of the iOS version, the underlying type of NSDate may change.
-            //Knowing that we have an error is enough.
-            XCTAssertTrue(error.localizedDescription.starts(with: "Event cannot be converted to JSON (Invalid type in JSON write"))
-        }
-        waitForExpectations(timeout: 1)
+        let data = SentrySerialization.data(withJSONObject: json)
         XCTAssertNil(data)
     }
 
@@ -198,9 +187,9 @@ class SentrySerializationTests: XCTestCase {
         let dict = SentrySession(releaseName: "1.0.0").serialize()
         let session = SentrySession(jsonObject: dict)!
         
-        let data = try SentrySerialization.data(with: session)
+        let data = SentrySerialization.data(with: session)
         
-        XCTAssertNotNil(SentrySerialization.session(with: data))
+        XCTAssertNotNil(SentrySerialization.session(with: data!))
     }
     
     func testSerializeSessionWithNoReleaseName() throws {
@@ -208,7 +197,7 @@ class SentrySerializationTests: XCTestCase {
         dict["attrs"] = nil // Remove release name
         let session = SentrySession(jsonObject: dict)!
         
-        let data = try SentrySerialization.data(with: session)
+        let data = SentrySerialization.data(with: session)!
         
         XCTAssertNil(SentrySerialization.session(with: data))
     }
@@ -217,7 +206,7 @@ class SentrySerializationTests: XCTestCase {
         let dict = SentrySession(releaseName: "").serialize()
         let session = SentrySession(jsonObject: dict)!
         
-        let data = try SentrySerialization.data(with: session)
+        let data = SentrySerialization.data(with: session)!
         
         XCTAssertNil(SentrySerialization.session(with: data))
     }
@@ -225,7 +214,7 @@ class SentrySerializationTests: XCTestCase {
     func testSerializeSessionWithGarbageInDict() throws {
         var dict = SentrySession(releaseName: "").serialize()
         dict["started"] = "20"
-        let data = try SentrySerialization.data(withJSONObject: dict)
+        let data = SentrySerialization.data(withJSONObject: dict)!
         
         XCTAssertNil(SentrySerialization.session(with: data))
     }
@@ -252,7 +241,7 @@ class SentrySerializationTests: XCTestCase {
     
     func testAppStateWithValidData_ReturnsValidAppState() throws {
         let appState = TestData.appState
-        let appStateData = try SentrySerialization.data(withJSONObject: appState.serialize())
+        let appStateData = SentrySerialization.data(withJSONObject: appState.serialize())!
         
         let actual = SentrySerialization.appState(with: appStateData)